Wall slide, movement, and textures
Added wall sliding and wall jumping, added a sprite to the jumper, added a background, and adjusted the colors of platforms. Made the first level and adjusted movement to my liking. Made a handful of prefabs too.
Next is a goal, a spawn point, and hazards (spikes). Also more levels. Figure out how to add text to backgrounds without UI stuff.