Added shifting platforms and spikes, and added more levels
I added platforms and spikes that can be active or inactive based on the state of the level. I also made more levels featuring this mechanic and an end screen. I handled a few other small bugs I came across.
The final thing I did was add a fully functioning build to the github repo.